requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jose==1.0.0
packaging==24.2
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.5
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
pypng==0.20220715.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
qrcode==7.4.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
